feat: move node related code into `environments.ts`
