+++ packages/core/e2e/edge-runtime/src/basic.test.ts
@@ -1,5 +1,5 @@
 describe("basic", () => {
-  it("import llamaindex", async () => {
+  it("import llamaindex should work", async () => {
     await import("llamaindex");
   });
 });
