--- packages/core/src/index.edge-light.ts
@@ -1,4 +1,0 @@
-console.warn("edge runtime is still under heavy development.");
-console.warn("please use the nodejs runtime for the full feature set.");
-
-export {};
