+++ packages/core/src/storage/FileSystem.edge-light.ts
@@ -1,0 +1,1 @@
+export * from "./FileSystem.core.js";
